+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Fixed
 - [ V2 ] Fix a sprints calc bug.
 
+### Added
+- [ V2 ] Dialog confirmation when story state is changed in expanded shape.
+
 ### Changed
 - [ V2 ] When story is accepted it has a different background color.